% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abPlot.R
\name{abPlot}
\alias{abPlot}
\title{AB plot of sender/receiver effects}
\usage{
abPlot(muEff, ylabel = "Sender Additive Effects")
}
\arguments{
\item{muEff}{A vector of sender additive effects}
}
\value{
A ggplot object
}
\description{
AB plot of sender/receiver effects
}
\author{
Shahryar Minhas
}
