% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_utils.R
\name{ame_memory_settings}
\alias{ame_memory_settings}
\title{Display memory usage information for AME models}
\usage{
ame_memory_settings(n_nodes, R = 2)
}
\arguments{
\item{n_nodes}{Number of nodes in network}

\item{R}{Rank of multiplicative effects (default: 2)}
}
\value{
Invisibly returns memory estimates
}
\description{
Shows estimated memory usage for networks of given size.
Memory optimization is now automatic, so this is informational only.
}
\details{
Memory optimization is now user-controlled:
\itemize{
\item Use remove_redundant=TRUE to remove EZ and UVPM matrices
\item Use use_sparse_matrices=TRUE for sparse networks
\item Both options can be combined for maximum memory savings
}
}
\author{
Cassy Dorff, Shahryar Minhas, Tosin Salau
}
