% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_utils.R
\name{ame_memory_usage}
\alias{ame_memory_usage}
\title{Calculate memory usage of AME model components}
\usage{
ame_memory_usage(fit, detailed = TRUE)
}
\arguments{
\item{fit}{Fitted AME model}

\item{detailed}{Logical; show detailed breakdown (default TRUE)}
}
\value{
Invisibly returns data.frame with memory usage
}
\description{
Calculate memory usage of AME model components
}
\author{
Cassy Dorff, Shahryar Minhas, Tosin Salau
}
