% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_wrappers.R
\name{array_to_list_opt}
\alias{array_to_list_opt}
\title{Optimized array_to_list with C++ backend}
\usage{
array_to_list_opt(arr, actorByYr, pdLabs)
}
\description{
Optimized array_to_list with C++ backend
}
\keyword{internal}
