% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_helpers.R
\name{bipartite_helpers}
\alias{bipartite_helpers}
\alias{init_bipartite_startvals}
\title{Bipartite network helper functions}
\usage{
init_bipartite_startvals(
  Y,
  family,
  nA,
  nB,
  RA,
  RB,
  T,
  Xlist = NULL,
  odmax = NULL
)
}
\description{
Bipartite network helper functions
}
\keyword{Initialize}
\keyword{bipartite}
\keyword{internal}
\keyword{starting}
\keyword{values}
