% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format.R
\name{check_format}
\alias{check_format}
\title{Check formatting of input objects into ame_repL function}
\usage{
check_format(Y, Xdyad, Xrow, Xcol)
}
\arguments{
\item{Y}{a T length list of n x n relational matrices, where T corresponds to the number of replicates (over time, for example). See
family below for various data types.}

\item{Xdyad}{a T length list of n x n x pd arrays of covariates}

\item{Xrow}{a T length list of n x pr matrices of nodal row covariates}

\item{Xcol}{a T length list of n x pc matrices of nodal column covariates}
}
\description{
Function used within ame_repL to ensure that input objects are formatted
correctly. Will stop function run if discrepancies found.
}
\author{
Shahryar Minhas
}
