% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/als_dynamic_beta.R
\name{coef.als_dynamic_beta}
\alias{coef.als_dynamic_beta}
\title{Extract beta path from a penalised-ALS object}
\usage{
\method{coef}{als_dynamic_beta}(object, ...)
}
\arguments{
\item{object}{An \code{als_dynamic_beta} object.}

\item{...}{Ignored.}
}
\value{
The \eqn{p \times T} matrix of estimates.
}
\description{
Extract beta path from a penalised-ALS object
}
