% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_parallel.R
\name{combine_ame_chains}
\alias{combine_ame_chains}
\title{Combine multiple AME chains}
\usage{
combine_ame_chains(chain_list, diagnostics = TRUE)
}
\arguments{
\item{chain_list}{List of ame fit objects from multiple chains}

\item{diagnostics}{Logical; whether to compute convergence diagnostics (default TRUE)}
}
\value{
A single ame object with combined chains
}
\description{
Combine multiple AME chains
}
\author{
Cassy Dorff, Shahryar Minhas, Tosin Salau
}
