% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_helpers.R
\name{compute_gof_bipartite}
\alias{compute_gof_bipartite}
\title{Compute GOF statistics for bipartite networks}
\usage{
compute_gof_bipartite(Y_obs, Y_sim, family)
}
\description{
Compute GOF statistics for bipartite networks
}
\keyword{internal}
