% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_XtX_Xty_bip_cpp}
\alias{compute_XtX_Xty_bip_cpp}
\title{Compute X'X and X'y for bipartite covariate regression}
\usage{
compute_XtX_Xty_bip_cpp(Xlist, resid, p)
}
\arguments{
\item{Xlist}{List of T arrays, each nA x nB x p}

\item{resid}{3D array of residuals nA x nB x T}

\item{p}{Number of covariates}
}
\value{
List with XtX (p x p) and Xty (p vector)
}
\description{
Replaces the O(T x p^2 x n^2) nested R loop for bipartite XtX and Xty
computation with a single C++ call.
}
