% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ame.R
\name{fitted.ame}
\alias{fitted.ame}
\title{Extract fitted values from AME model}
\usage{
\method{fitted}{ame}(object, ...)
}
\arguments{
\item{object}{Fitted AME model}

\item{...}{Additional arguments}
}
\value{
Matrix of fitted values
}
\description{
Extract fitted values from AME model
}
\author{
Cassy Dorff, Shahryar Minhas, Tosin Salau
}
