% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_als.R
\name{fitted.ame_als}
\alias{fitted.ame_als}
\title{Extract fitted values from a fast AME fit}
\usage{
\method{fitted}{ame_als}(object, ...)
}
\arguments{
\item{object}{an \code{ame_als} object.}

\item{...}{ignored.}
}
\value{
A matrix or list of matrices of fitted values.
}
\description{
Returns response-scale fitted values: a single matrix for a cross-sectional
(\code{ame_als}) fit, or a list of matrices for a longitudinal
(\code{lame_als}) fit.
}
