% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ame.R
\name{fitted.lame}
\alias{fitted.lame}
\title{Extract fitted values from LAME model}
\usage{
\method{fitted}{lame}(object, ...)
}
\arguments{
\item{object}{Fitted LAME model}

\item{...}{Additional arguments}
}
\value{
List of fitted value matrices (one per time point)
}
\description{
Extract fitted values from LAME model
}
