% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design_rep.R
\name{get_design_rep}
\alias{get_design_rep}
\title{Create design array for replicate data}
\usage{
get_design_rep(Y, Xdyad, Xrow, Xcol, actorSet, intercept, n, N, pr, pc, pd)
}
\arguments{
\item{Y}{dependent variable in array format}

\item{Xdyad}{dyadic covariates in array format}

\item{Xrow}{sender covariates in array format}

\item{Xcol}{receiver covariates in array format}

\item{actorSet}{vector of actors}

\item{intercept}{logical indicating whether to include intercept}

\item{n}{number of actors}

\item{N}{number of replicates}

\item{pr}{number of receiver covariates}

\item{pc}{number of sender covariates}

\item{pd}{number of dyadic covariates}
}
\value{
returns list of design array values necessary for ame_repL
}
\description{
Create design array for replicate data
}
\author{
Shahryar Minhas
}
