% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCirc.R
\name{ggCirc}
\alias{ggCirc}
\title{Plot circular network with ggplot2}
\usage{
ggCirc(
  Y,
  U = NULL,
  V = NULL,
  row.names = rownames(Y),
  col.names = colnames(Y),
  vscale = 0.6,
  prange = c(2, 5),
  lcol = "gray85",
  ltype = "dotted",
  lsize = 0.5,
  force = 2,
  maxIter = 3000,
  showActLinks = FALSE,
  geomLabel = TRUE,
  geomText = FALSE,
  geomPoint = TRUE,
  ...
)
}
\arguments{
\item{Y}{(matrix) m by n relational matrix.}

\item{U}{(matrix) m by 2 matrix of row factors of Y.}

\item{V}{(matrix) n by 2 matrix of column factors of Y.}

\item{row.names}{(character vector) names of the row objects.}

\item{col.names}{(character vector) names of the columns objects.}

\item{vscale}{Scaling factor for the V matrix}

\item{prange}{Range for point size}

\item{lcol}{Color for the lines}

\item{ltype}{Line type for the lines}

\item{lsize}{Line size for the lines}

\item{force}{Force parameter for ggrepel}

\item{maxIter}{Maximum iterations for ggrepel}

\item{showActLinks}{Logical indicating whether to show active links}

\item{geomLabel}{Logical indicating whether to use geom_label}

\item{geomText}{Logical indicating whether to use geom_text}

\item{geomPoint}{Logical indicating whether to use geom_point}
}
\value{
A ggplot object representing the circular network
}
\description{
Plot circular network with ggplot2
}
\author{
Shahryar Minhas
}
