% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_dynamic_ab_cpp}
\alias{init_dynamic_ab_cpp}
\title{Initialize dynamic additive effects with AR(1) structure}
\usage{
init_dynamic_ab_cpp(n, T, rho_ab, sigma_ab, mean_a = 0, mean_b = 0)
}
\arguments{
\item{n}{Number of actors}

\item{T}{Number of time points}

\item{rho_ab}{AR(1) parameter}

\item{sigma_ab}{Innovation standard deviation}

\item{mean_a}{Mean for row effects}

\item{mean_b}{Mean for column effects}
}
\value{
List with initialized a and b matrices
}
\description{
Initialize dynamic additive effects with AR(1) structure
}
