% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_array.R
\name{list_to_array}
\alias{list_to_array}
\title{Convert list to array}
\usage{
list_to_array(actors, Y, Xdyad, Xrow, Xcol)
}
\arguments{
\item{actors}{vector of actors}

\item{Y}{dv in list format}

\item{Xdyad}{dyadic covariates in list format}

\item{Xrow}{sender covariates in list format}

\item{Xcol}{receiver covariates in list format}
}
\value{
transforms Y, Xdyad, Xrow, and Xcol to arrays
}
\description{
Convert list to array
}
\author{
Shahryar Minhas
}
