% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_wrappers.R
\name{llsrmRho_opt}
\alias{llsrmRho_opt}
\title{Optimized llsrmRho with C++ backend}
\usage{
llsrmRho_opt(Y, Sab, rhos, s2 = 1)
}
\description{
Optimized llsrmRho with C++ backend
}
\keyword{internal}
