% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison_methods.R
\name{nobs.ame}
\alias{nobs.ame}
\alias{nobs.lame}
\title{Number of observed dyads in an AME / LAME fit}
\usage{
\method{nobs}{ame}(object, ...)

\method{nobs}{lame}(object, ...)
}
\arguments{
\item{object}{an \code{ame} or \code{lame} fit.}

\item{...}{ignored.}
}
\value{
Integer: number of observed dyads in \code{Y}.
}
\description{
Number of finite (non-missing) cells in \code{Y}, summed over time slices
for longitudinal fits. Useful as a denominator for sample-size reporting
and as the basis for AIC/BIC \emph{if} you also have a log-likelihood
(which \code{ame()} / \code{lame()} do not directly expose; see
\code{\link{logLik.ame}}).
}
