% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_utils.R
\name{posterior_quantiles}
\alias{posterior_quantiles}
\title{Extract posterior quantiles for model components}
\usage{
posterior_quantiles(
  fit,
  component = c("beta", "UV", "ab"),
  probs = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{fit}{Fitted ame model object}

\item{component}{Character; which component: "beta", "UV", "ab"}

\item{probs}{Numeric vector of probabilities for quantiles}
}
\value{
Matrix or array of posterior quantiles
}
\description{
Extract posterior quantiles for model components
}
\author{
Cassy Dorff, Shahryar Minhas, Tosin Salau
}
