% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomputeX.R
\name{precomputeX}
\alias{precomputeX}
\title{Precompute design matrix statistics}
\usage{
precomputeX(X)
}
\arguments{
\item{X}{n x n x p covariate array}
}
\value{
The input array with precomputed statistics as attributes:
\item{Xr}{row sums}
\item{Xc}{column sums}
\item{mX}{matricized version}
\item{mXt}{dyad-transposed matricized version}
\item{XX}{regression sums of squares}
\item{XXt}{crossproduct sums of squares}
}
\description{
Precomputes summary statistics of the design array X to speed up
MCMC calculations in the AME model
}
\author{
Peter Hoff
}
