% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sim.R
\name{print.ame.sim}
\alias{print.ame.sim}
\title{Print methods for AME and LAME simulation objects}
\usage{
\method{print}{ame.sim}(x, ...)
}
\arguments{
\item{x}{simulation object of class "ame.sim" or "lame.sim"}

\item{...}{additional arguments (not used)}
}
\value{
the simulation object invisibly
}
\description{
Print methods for AME and LAME simulation objects
}
\author{
Shahryar Minhas
}
