% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_als.R
\name{print.ame_als}
\alias{print.ame_als}
\title{Print an ame_als object}
\usage{
\method{print}{ame_als}(x, digits = 4, ...)
}
\arguments{
\item{x}{an \code{ame_als} object.}

\item{digits}{number of digits to display.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an ame_als object
}
