% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lame_multi.R
\name{print.lame_multi}
\alias{print.lame_multi}
\title{Print method for lame_multi}
\usage{
\method{print}{lame_multi}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{lame_multi} object.}

\item{digits}{Number of significant digits.}

\item{...}{Ignored.}
}
\description{
Print method for lame_multi
}
