% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo_lame.R
\name{print.lfo_lame}
\alias{print.lfo_lame}
\title{Print method for lfo() results}
\usage{
\method{print}{lfo_lame}(x, ...)
}
\arguments{
\item{x}{A \code{lfo_lame} object.}

\item{...}{Ignored.}
}
\description{
Print method for lfo() results
}
