% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_actor_slopes.R
\name{print.per_actor_slopes}
\alias{print.per_actor_slopes}
\title{Print method for per_actor_slopes}
\usage{
\method{print}{per_actor_slopes}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{per_actor_slopes} result.}

\item{digits}{Number of significant digits.}

\item{...}{Ignored.}
}
\description{
Print method for per_actor_slopes
}
