% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ame.R
\name{print.summary.ame}
\alias{print.summary.ame}
\title{Print method for summary.ame objects}
\usage{
\method{print}{summary.ame}(x, ...)
}
\arguments{
\item{x}{a summary.ame object}

\item{...}{additional arguments (not used)}
}
\value{
the summary.ame object invisibly
}
\description{
Print method for summary.ame objects
}
