% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbeta_ab_bip_gibbs_cpp}
\alias{rbeta_ab_bip_gibbs_cpp}
\title{Full bipartite Gibbs update for beta, a, b}
\usage{
rbeta_ab_bip_gibbs_cpp(
  Z,
  Xlist,
  UV_eff,
  a_current,
  b_current,
  s2,
  g_prior,
  va,
  vb,
  rvar,
  cvar
)
}
\arguments{
\item{Z}{3D array (nA x nB x T)}

\item{Xlist}{List of T arrays (nA x nB x p)}

\item{UV_eff}{nA x nB matrix (U\emph{G}V' or U*V')}

\item{a_current}{Current row effects (length nA)}

\item{b_current}{Current column effects (length nB)}

\item{s2}{Dyadic variance}

\item{g_prior}{G-prior parameter}

\item{va}{Row effect variance (diagonal element of Sab)}

\item{vb}{Column effect variance (diagonal element of Sab)}

\item{rvar}{Whether to update row effects}

\item{cvar}{Whether to update column effects}
}
\value{
List with beta, a, b
}
\description{
Single C++ function replacing the bipartite beta/a/b update block in lame.R
}
