% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{rbeta_ab_fc_opt}
\alias{rbeta_ab_fc_opt}
\title{Optimized rbeta_ab_fc with C++ backend}
\usage{
rbeta_ab_fc_opt(
  Z,
  Sab,
  rho,
  X = NULL,
  s2 = 1,
  offset = 0,
  iV0 = NULL,
  m0 = NULL,
  g = length(Z)
)
}
\description{
Optimized rbeta_ab_fc with C++ backend
}
\keyword{internal}
