% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ame.R
\name{residuals.lame}
\alias{residuals.lame}
\title{Extract residuals from LAME model}
\usage{
\method{residuals}{lame}(object, type = c("response", "pearson"), ...)
}
\arguments{
\item{object}{Fitted LAME model}

\item{type}{Type of residuals ("response" or "pearson")}

\item{...}{Additional arguments}
}
\value{
List of residual matrices (one per time point)
}
\description{
Extract residuals from LAME model
}
