% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_wrappers.R
\name{rrho_fc_opt}
\alias{rrho_fc_opt}
\title{Optimized rrho_fc with C++ backend}
\usage{
rrho_fc_opt(Z, Sab, s2 = 1, offset = 0, ngp = 100, asp = TRUE)
}
\description{
Optimized rrho_fc with C++ backend
}
\keyword{internal}
