% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rUV_dynamic_fc_cpp}
\alias{rUV_dynamic_fc_cpp}
\title{Update dynamic latent positions using AR(1) process}
\usage{
rUV_dynamic_fc_cpp(
  U_current,
  V_current,
  ET,
  rho_uv,
  sigma_uv,
  s2,
  shrink,
  symmetric
)
}
\arguments{
\item{U_current}{Current 3D array of U positions (n x R x T)}

\item{V_current}{Current 3D array of V positions (n x R x T)}

\item{ET}{3D array of residuals (n x n x T)}

\item{rho_uv}{AR(1) autoregressive parameter for latent positions}

\item{sigma_uv}{Innovation standard deviation for latent positions}

\item{s2}{Dyadic variance}

\item{shrink}{Whether to apply shrinkage}

\item{symmetric}{Whether network is symmetric}
}
\value{
List with updated U and V arrays
}
\description{
Update dynamic latent positions using AR(1) process
}
