% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rZ_bin_bip_batch_cpp}
\alias{rZ_bin_bip_batch_cpp}
\title{Batch binary Z sampling across all time periods (bipartite, rho=0)}
\usage{
rZ_bin_bip_batch_cpp(Z, EZ, Y)
}
\arguments{
\item{Z}{3D array of latent values (nA x nB x T)}

\item{EZ}{3D array of expected values (nA x nB x T)}

\item{Y}{3D array of observed values (nA x nB x T)}
}
\value{
Updated Z array
}
\description{
Vectorized probit update for bipartite binary networks without dyadic correlation.
}
