% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rZ_pois_fc.R
\name{rZ_pois_fc}
\alias{rZ_pois_fc}
\title{Gibbs update for latent variable in a Poisson AME model}
\usage{
rZ_pois_fc(Z, EZ, rho, s2, Y)
}
\arguments{
\item{Z}{a square matrix, the current value of the latent variable}

\item{EZ}{expected value of Z (regression effects + random effects)}

\item{rho}{dyadic correlation}

\item{s2}{dyadic variance (overdispersion parameter)}

\item{Y}{square relational matrix of observed counts}
}
\value{
updated value of Z
}
\description{
Updates the latent variable Z in a Poisson AME model using a
Metropolis-Hastings step. The model assumes y_\{i,j\} ~ Poisson(exp(z_\{i,j\}))
where z_\{i,j\} is the latent variable representing the log mean.
}
\author{
Cassy Dorff, Shahryar Minhas, Tosin Salau
}
