% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_helpers.R
\name{sample_ab_bipartite}
\alias{sample_ab_bipartite}
\title{Sample additive effects for bipartite networks}
\usage{
sample_ab_bipartite(Z, EZ_without_ab, sigma2_a, sigma2_b, s2)
}
\description{
Sample additive effects for bipartite networks
}
\keyword{internal}
