% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_sigma_ab_cpp}
\alias{sample_sigma_ab_cpp}
\title{Sample innovation variance for dynamic additive effects}
\usage{
sample_sigma_ab_cpp(
  a_mat,
  b_mat,
  rho_ab,
  symmetric,
  prior_shape = 2,
  prior_scale = 1
)
}
\arguments{
\item{a_mat}{Matrix of row effects (n x T)}

\item{b_mat}{Matrix of column effects (n x T)}

\item{rho_ab}{AR(1) parameter}

\item{symmetric}{Whether the network is symmetric}

\item{prior_shape}{Shape parameter for inverse gamma prior}

\item{prior_scale}{Scale parameter for inverse gamma prior}
}
\value{
Updated sigma_ab value
}
\description{
Sample innovation variance for dynamic additive effects
}
