% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simY_pois.R
\name{simY_pois}
\alias{simY_pois}
\title{Simulate a Poisson relational matrix}
\usage{
simY_pois(EZ)
}
\arguments{
\item{EZ}{square matrix giving the log expected values of the relational matrix}
}
\value{
a square matrix of counts
}
\description{
Simulates a relational matrix from a Poisson distribution given
the log mean values
}
\author{
Cassy Dorff, Shahryar Minhas, Tosin Salau
}
