% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sim.R
\name{summary.ame.sim}
\alias{summary.ame.sim}
\title{Summary method for AME simulations}
\usage{
\method{summary}{ame.sim}(object, ...)
}
\arguments{
\item{object}{simulation object of class "ame.sim"}

\item{...}{additional arguments (not used)}
}
\value{
summary statistics invisibly
}
\description{
Summary method for AME simulations
}
