% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_als.R
\name{summary.ame_als}
\alias{summary.ame_als}
\title{Summarize an ame_als object}
\usage{
\method{summary}{ame_als}(object, ...)
}
\arguments{
\item{object}{an \code{ame_als} object.}

\item{...}{ignored.}
}
\value{
A list of class \code{"summary.ame_als"}, printed as a table.
}
\description{
Summarize an ame_als object
}
