% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sim.R
\name{summary.lame.sim}
\alias{summary.lame.sim}
\title{Summary method for LAME simulations}
\usage{
\method{summary}{lame.sim}(object, ...)
}
\arguments{
\item{object}{simulation object of class "lame.sim"}

\item{...}{additional arguments (not used)}
}
\value{
summary statistics
}
\description{
Summary method for LAME simulations
}
