% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_helpers.R
\name{update_variances_bipartite}
\alias{update_variances_bipartite}
\title{Update variance parameters for bipartite}
\usage{
update_variances_bipartite(
  a,
  b,
  eta0_a = 2,
  eta0_b = 2,
  Sab0_aa = 1,
  Sab0_bb = 1
)
}
\description{
Update variance parameters for bipartite
}
\keyword{internal}
