% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ame.R
\name{vcov.ame}
\alias{vcov.ame}
\alias{vcov.lame}
\title{Posterior covariance of AME model coefficients}
\usage{
\method{vcov}{ame}(object, ...)

\method{vcov}{lame}(object, ...)
}
\arguments{
\item{object}{fitted AME model (class "ame")}

\item{...}{additional arguments (ignored)}
}
\value{
p x p covariance matrix of posterior BETA draws
}
\description{
Returns the posterior covariance matrix of regression coefficients.
}
