% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Xbeta_bip_cpp}
\alias{Xbeta_bip_cpp}
\title{Compute Xbeta product for bipartite networks}
\usage{
Xbeta_bip_cpp(X, beta)
}
\arguments{
\item{X}{3D array (nA x nB x p) of covariates for one time period}

\item{beta}{Coefficient vector of length p}
}
\value{
nA x nB matrix
}
\description{
Computes sum_k beta_k * X_k for a single time slice
}
