% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Xcol}
\alias{Xcol}
\title{Column covariates}
\format{
A list of matrices of column covariates
}
\description{
Column covariates loaded by \code{data(vignette_data)}.
See \code{\link{vignette_data}} for full details.
}
\seealso{
\code{\link{vignette_data}}
}
\keyword{datasets}
