% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Xdyad}
\alias{Xdyad}
\title{Dyadic covariates}
\format{
A list of arrays of dyadic covariates
}
\description{
Dyadic covariates loaded by \code{data(vignette_data)}.
See \code{\link{vignette_data}} for full details.
}
\seealso{
\code{\link{vignette_data}}
}
\keyword{datasets}
