% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame.R
\name{ame}
\alias{ame}
\title{AME model fitting routine}
\usage{
ame(Y, Xdyad=NULL, Xrow=NULL, Xcol=NULL, rvar = !(family=="rrl") ,
cvar = TRUE,  dcor = !symmetric, nvar=TRUE, R = 0, family="normal",
intercept=!is.element(family,c("rrl","ordinal")),
symmetric=FALSE,
odmax=rep(max(apply(Y>0,1,sum,na.rm=TRUE)),nrow(Y)), 
prior=list(), g=NA,
seed = 6886, nscan = 10000, burn = 500, odens = 25, 
plot=TRUE, print = TRUE, gof=TRUE, model.name=NULL)
}
\arguments{
\item{Y}{an n x n square relational matrix of relations. See family below for
various data types.}

\item{Xdyad}{an n x n x pd array of covariates}

\item{Xrow}{an n x pr matrix of nodal row covariates}

\item{Xcol}{an n x pc matrix of nodal column covariates}

\item{rvar}{logical: fit row random effects (asymmetric case)?}

\item{cvar}{logical: fit column random effects (asymmetric case)?}

\item{dcor}{logical: fit a dyadic correlation (asymmetric case)?}

\item{nvar}{logical: fit nodal random effects (symmetric case)?}

\item{R}{integer: dimension of the multiplicative effects (can be zero)}

\item{family}{character: one of "normal","tobit","binary","ordinal","cbin","frn","rrl","poisson" - see
the details below}

\item{intercept}{logical: fit model with an intercept?}

\item{symmetric}{logical: Is the sociomatrix symmetric by design?}

\item{odmax}{a scalar integer or vector of length n giving the maximum
number of nominations that each node may make - used for "frn" and "cbin"
families}

\item{prior}{a list containing hyperparameters for the prior distributions}

\item{g}{optional scalar or vector of length dim(X)\[3\] for g-prior}

\item{seed}{random seed}

\item{nscan}{number of iterations of the Markov chain (beyond burn-in)}

\item{burn}{burn in for the Markov chain}

\item{odens}{output density for the Markov chain}

\item{plot}{logical: plot results while running?}

\item{print}{logical: print results while running?}

\item{gof}{logical: calculate goodness of fit statistics?}

\item{model.name}{optional string for model selection output}
}
\value{
\item{BETA}{posterior samples of regression coefficients}
\item{VC}{posterior samples of the variance parameters}
\item{APM}{posterior mean of additive row effects a} \item{BPM}{posterior
mean of additive column effects b} \item{U}{posterior mean of multiplicative
row effects u} \item{V}{posterior mean of multiplicative column effects v (asymmetric case)}
\item{UVPM}{posterior mean of UV (asymmetric case)}
\item{ULUPM}{posterior mean of ULU (symmetric case)}
\item{L}{posterior mean of L (symmetric case)}
\item{EZ}{estimate of expectation of Z
matrix} \item{YPM}{posterior mean of Y (for imputing missing values)}
\item{GOF}{observed (first row) and posterior predictive (remaining rows)
values of four goodness-of-fit statistics}
\item{AIC}{Akaike Information Criterion (if model.name provided)}
\item{BIC}{Bayesian Information Criterion (if model.name provided)}
\item{model.name}{Name of the model (if provided)}
}
\description{
An MCMC routine providing a fit to an additive and multiplicative effects
(AME) regression model to relational data of various types
}
\details{
This command provides posterior inference for parameters in AME models of
relational data, assuming one of eight possible data types/models:

"normal": A normal AME model.

"tobit": A tobit AME model for censored continuous data. Values are censored
at zero, appropriate for non-negative continuous relational data.

"binary": A binary probit AME model.

"ordinal": An ordinal probit AME model. An intercept is not identifiable in this
model.

"cbin": An AME model for censored binary data.  The value of 'odmax'
specifies the maximum number of links each row may have.

"frn": An AME model for fixed rank nomination networks. A higher value of
the rank indicates a stronger relationship. The value of 'odmax' specifies
the maximum number of links each row may have.

"rrl": An AME model based on the row ranks. This is appropriate if the
relationships across rows are not directly comparable in terms of scale. An
intercept, row random effects and row regression effects are not estimable
for this model.

"poisson": An overdispersed Poisson AME model for count data. The latent
variable represents the log mean of the Poisson distribution.
}
\examples{
\dontrun{
data(YX_bin) 
fit<-ame(YX_bin$Y,YX_bin$X,burn=10,nscan=10,odens=1,family="binary")
# Note: you should run the Markov chain much longer in practice
} 
 
}
\author{
Peter Hoff
}
