% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_parallel.R
\name{compute_mcmc_diagnostics}
\alias{compute_mcmc_diagnostics}
\title{Compute MCMC convergence diagnostics for multiple chains}
\usage{
compute_mcmc_diagnostics(chain_list)
}
\arguments{
\item{chain_list}{List of ame fit objects from multiple chains}
}
\value{
List containing convergence diagnostics
}
\description{
Compute MCMC convergence diagnostics for multiple chains
}
\author{
Cassy Dorff, Shahryar Minhas, Tosin Salau
}
