% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_object.R
\name{get_fit_object}
\alias{get_fit_object}
\title{Get fitted object from MCMC results}
\usage{
get_fit_object(
  APS,
  BPS,
  UVPS,
  YPS,
  BETA,
  VC,
  GOF,
  Xlist,
  actorByYr,
  startVals,
  symmetric,
  tryErrorChecks,
  AIC = NA,
  BIC = NA,
  model.name = NULL,
  U = NULL,
  V = NULL,
  dynamic_uv = FALSE,
  dynamic_ab = FALSE,
  bip = FALSE,
  rho_ab = NULL,
  rho_uv = NULL
)
}
\arguments{
\item{APS}{summed additive sender random effects (or matrix for dynamic)}

\item{BPS}{summed additive receiver random effects (or matrix for dynamic)}

\item{UVPS}{summed multiplicative random effects}

\item{YPS}{summed Y posterior predictive values}

\item{BETA}{Matrix of draws for regression coefficient estimates}

\item{VC}{Matrix of draws for variance estimates}

\item{GOF}{Matrix of draws for goodness of fit calculations}

\item{Xlist}{List based version of design array}

\item{actorByYr}{List of actors by time point}

\item{startVals}{startVals for future model run}

\item{symmetric}{logical indicating whether model is symmetric}

\item{tryErrorChecks}{list with counts of MCMC errors}

\item{AIC}{Akaike Information Criterion (optional)}

\item{BIC}{Bayesian Information Criterion (optional)}

\item{model.name}{Name of the model (optional)}

\item{U}{Latent sender positions (optional, for dynamic UV)}

\item{V}{Latent receiver positions (optional, for dynamic UV)}

\item{dynamic_uv}{logical indicating whether UV effects are dynamic}

\item{dynamic_ab}{logical indicating whether additive effects are dynamic}
}
\value{
Fitted AME object
}
\description{
Get fitted object from MCMC results
}
\author{
Peter Hoff, Shahryar Minhas
}
