% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_vals.R
\name{get_start_vals}
\alias{get_start_vals}
\title{Get fitted object from MCMC results}
\usage{
get_start_vals(start_vals, Y, family, xP, rvar, cvar, R, odmax = NULL)
}
\arguments{
\item{start_vals}{List object that is null or contains
starting values}

\item{Y}{dependent variable in array format}

\item{family}{character vector (e.g. 'bin', 'nrm') specifying
family type}

\item{xP}{number of exogenous covariates}

\item{rvar}{logical indicating whether to include sender random
effects}

\item{cvar}{logical indicating whether to include receiver random
effects}

\item{R}{Number of dimensions for multiplicative effects}

\item{odmax}{vector of maximum ranks for cbin/frn families (optional)}
}
\value{
List of starting values for MCMC
}
\description{
Get fitted object from MCMC results
}
\author{
Shahryar Minhas
}
