% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_helpers.R
\name{list_to_array_bipartite}
\alias{list_to_array_bipartite}
\title{Convert bipartite list data to array format}
\usage{
list_to_array_bipartite(
  rowActorSet,
  colActorSet,
  Y_list,
  Xdyad = NULL,
  Xrow = NULL,
  Xcol = NULL
)
}
\description{
Convert bipartite list data to array format
}
\keyword{internal}
