% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ame.R
\name{print.ame}
\alias{print.ame}
\title{Print method for AME objects}
\usage{
\method{print}{ame}(x, ...)
}
\arguments{
\item{x}{an object of class "ame"}

\item{...}{additional arguments (not used)}
}
\value{
the ame object invisibly
}
\description{
Provides a concise print output for fitted AME models
}
\author{
Peter Hoff, Cassy Dorff, Shahryar Minhas
}
