% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lame.R
\name{print.lame}
\alias{print.lame}
\title{Print method for LAME objects}
\usage{
\method{print}{lame}(x, ...)
}
\arguments{
\item{x}{an object of class "lame"}

\item{...}{additional arguments (not used)}
}
\value{
the lame object invisibly
}
\description{
Provides a concise print output for fitted LAME models
}
\author{
Peter Hoff, Cassy Dorff, Shahryar Minhas
}
