% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ame.R
\name{residuals.ame}
\alias{residuals.ame}
\title{Extract residuals from AME model}
\usage{
\method{residuals}{ame}(object, type = c("response", "pearson"), ...)
}
\arguments{
\item{object}{Fitted AME model}

\item{type}{Type of residuals ("response" or "pearson")}

\item{...}{Additional arguments}
}
\value{
Matrix of residuals
}
\description{
Extract residuals from AME model
}
\author{
Cassy Dorff, Shahryar Minhas, Tosin Salau
}
