% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_rho_ab_cpp}
\alias{sample_rho_ab_cpp}
\title{Sample AR(1) parameter for dynamic additive effects}
\usage{
sample_rho_ab_cpp(a_mat, b_mat, sigma_ab, rho_current, symmetric)
}
\arguments{
\item{a_mat}{Matrix of row effects (n x T)}

\item{b_mat}{Matrix of column effects (n x T)}

\item{sigma_ab}{Innovation standard deviation}

\item{rho_current}{Current value of rho}

\item{symmetric}{Whether the network is symmetric}
}
\value{
Updated rho value
}
\description{
Uses Metropolis-Hastings to sample rho_ab given the time series of effects
}
