% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_rho_uv}
\alias{sample_rho_uv}
\title{Sample AR(1) parameter for dynamic latent factors}
\usage{
sample_rho_uv(U_cube, V_cube, sigma_uv, rho_current, symmetric)
}
\arguments{
\item{U_cube}{3D array of U positions (n x R x T)}

\item{V_cube}{3D array of V positions (n x R x T)}

\item{sigma_uv}{Innovation standard deviation}

\item{rho_current}{Current value of rho}

\item{symmetric}{Whether network is symmetric}
}
\value{
Updated rho value
}
\description{
Sample AR(1) parameter for dynamic latent factors
}
